+++ projects/plugins/jetpack/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ### This is a list detailing changes for all Jetpack releases.
 
+## 12.5-a.7 - 2023-08-09
+### Enhancements
+- AI Extension: fix AI Assistant bar position in the Top toolbar mode. [#32374]
+- AI Extension: update view of AI assistant bar on mobile. [#32377]
+- SSO: enable "Match by Email" by default. [#32370]
+
 ## 12.5-a.5 - 2023-08-09
 ### Enhancements
 - AI Assistant: change tracked event to English instead of translated value and track language and tone change [#32291]
@@ -16,7 +22,7 @@
 ### Other changes <!-- Non-user-facing changes go here. This section will not be copied to readme.txt. -->
 - Blocks: avoid hardcoding the version number when creating a new block via the CLI command [#32110]
 - AI Control: clean up props [#32360]
-- AI Extension: add UpgradePrompt if upgrade is required. [#32326] 
+- AI Extension: add UpgradePrompt if upgrade is required. [#32326]
 - AI Plugin: Add upgrade path. [#32327]
 - Move out of api.js [#32273]
 - Updated package dependencies. [#32307]